+++ app/src/test/java/com/globant/counter/unitTest/MainActivityTest.kt
@@ -1,4 +1,4 @@
-package com.globant.counter
+package com.globant.counter.unitTest
 
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions.click
@@ -8,6 +8,8 @@ import androidx.test.espresso.matcher.ViewMatchers.withText
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.LargeTest
+import com.globant.counter.MainActivity
+import com.globant.counter.R
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
